Make your own change! Homework
Added my own image as the background. It is a file named
justkeepcoding.jpg which replaces the gray background on the original
snap. I also added a new block called random but I had some difficult
in locating all of the locations to make it fully functional. I also
changed a few colors.